intellij_idea//2022_1//13.07

+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleService.java
@@ -5,11 +5,11 @@ import ru.kata.spring.boot_security.demo.model.Role;
 import java.util.List;
 
 public interface RoleService {
-    Role getRoleById(int id);
+    Role getRoleById(long id);
     Role getRoleByRoleName(String name);
     void save(Role role);
     void update( Role updatedRole);
-    void delete(int id);
+    void delete(long id);
     List<Role> getDemandedRoles();
